DATA-1525 GTFSRT Externalization
DATA-1525 last

DATA-1525 last
