--- src/main/java/no/rutebanken/anshar/gtfsrt/mappers/VehiclePositionMapper.java
@@ -1,223 +1,0 @@
-package no.rutebanken.anshar.gtfsrt.mappers;
-
-
-import com.google.protobuf.Timestamp;
-import com.google.transit.realtime.GtfsRealtime;
-import no.rutebanken.anshar.gtfsrt.mappers.utils.ElementUtils;
-import no.rutebanken.anshar.routes.mapping.StopTimesService;
-import org.apache.commons.collections4.CollectionUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Component;
-import org.w3c.dom.Element;
-import uk.org.siri.siri20.*;
-
-import java.math.BigDecimal;
-import java.math.BigInteger;
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-import java.util.Collections;
-import java.util.Date;
-import java.util.List;
-
-
-/***
- * Utility class to convert vehiclePosition (GTFS RT) to vehicleActivity (SIRI)
- */
-@Component
-public class VehiclePositionMapper {
-
-    private static final DateFormat gtfsRtDateFormat = new SimpleDateFormat("yyyyMMdd HH:mm:ss");
-    private static final Logger logger = LoggerFactory.getLogger(VehiclePositionMapper.class);
-
-    private final StopTimesService stopTimesService;
-
-    public VehiclePositionMapper(StopTimesService stopTimesService) {
-        this.stopTimesService = stopTimesService;
-    }
-
-    private static void mapCongestion(VehicleActivityStructure.MonitoredVehicleJourney monitoredVehicleJourney, GtfsRealtime.VehiclePosition vehiclePosition) {
-        if (vehiclePosition.getCongestionLevel() != null) {
-            monitoredVehicleJourney.setInCongestion(GtfsRealtime.VehiclePosition.CongestionLevel.CONGESTION.equals(vehiclePosition.getCongestionLevel()));
-        }
-    }
-
-    private static void mapOccupancy(VehicleActivityStructure.MonitoredVehicleJourney monitoredVehicleJourney, GtfsRealtime.VehiclePosition vehiclePosition) {
-        if (vehiclePosition.getOccupancyStatus() != null) {
-            switch (vehiclePosition.getOccupancyStatus()) {
-                case FULL:
-                    monitoredVehicleJourney.setOccupancy(OccupancyEnumeration.FULL);
-                    break;
-                case STANDING_ROOM_ONLY:
-                    monitoredVehicleJourney.setOccupancy(OccupancyEnumeration.STANDING_AVAILABLE);
-                    break;
-                case FEW_SEATS_AVAILABLE:
-                    monitoredVehicleJourney.setOccupancy(OccupancyEnumeration.SEATS_AVAILABLE);
-                    break;
-            }
-        }
-    }
-
-    private static void mapStatus(VehicleActivityStructure activity, VehicleActivityStructure.MonitoredVehicleJourney monitoredVehicleJourney, GtfsRealtime.VehiclePosition vehiclePosition) {
-
-        if (GtfsRealtime.VehiclePosition.VehicleStopStatus.STOPPED_AT.equals(vehiclePosition.getCurrentStatus())) {
-            MonitoredCallStructure monitoredCallStruct = new MonitoredCallStructure();
-            LocationStructure locStruct = new LocationStructure();
-            locStruct.setLongitude(BigDecimal.valueOf(vehiclePosition.getPosition().getLongitude()));
-            locStruct.setLatitude(BigDecimal.valueOf(vehiclePosition.getPosition().getLatitude()));
-            monitoredCallStruct.setVehicleLocationAtStop(locStruct);
-
-            if (StringUtils.isNotEmpty(vehiclePosition.getStopId())) {
-                StopPointRef stopPointRef = new StopPointRef();
-                stopPointRef.setValue(vehiclePosition.getStopId());
-                monitoredCallStruct.setStopPointRef(stopPointRef);
-            }
-            monitoredCallStruct.setOrder(BigInteger.valueOf(vehiclePosition.getCurrentStopSequence()));
-            monitoredVehicleJourney.setMonitoredCall(monitoredCallStruct);
-
-        } else if (GtfsRealtime.VehiclePosition.VehicleStopStatus.INCOMING_AT.equals(vehiclePosition.getCurrentStatus())) {
-            // Arbitrary behaviour : "incoming at" status set to 90% progress
-            ProgressBetweenStopsStructure progressBetStruct = new ProgressBetweenStopsStructure();
-            progressBetStruct.setPercentage(BigDecimal.valueOf(90));
-            activity.setProgressBetweenStops(progressBetStruct);
-        }
-    }
-
-    private static void mapPosition(VehicleActivityStructure.MonitoredVehicleJourney monitoredVehicleJourney, GtfsRealtime.Position position) {
-        if (position == null)
-            return;
-
-        LocationStructure locationStructure = new LocationStructure();
-        locationStructure.setLatitude(BigDecimal.valueOf(position.getLatitude()));
-        locationStructure.setLongitude(BigDecimal.valueOf(position.getLongitude()));
-        monitoredVehicleJourney.setBearing(position.getBearing());
-        monitoredVehicleJourney.setVelocity(BigInteger.valueOf((long) position.getSpeed()));
-        monitoredVehicleJourney.setVehicleLocation(locationStructure);
-    }
-
-    private static void mapRecordedAtTime(VehicleActivityStructure activity, GtfsRealtime.VehiclePosition vehiclePosition) {
-        if (vehiclePosition.getTimestamp() != 0) {
-            ZonedDateTime timestamp = ZonedDateTime.ofInstant(Instant.ofEpochSecond(vehiclePosition.getTimestamp()), ZoneId.systemDefault());
-            activity.setRecordedAtTime(timestamp);
-        }
-    }
-
-    private static void mapTridData(VehicleActivityStructure.MonitoredVehicleJourney monitoredVehicleJourney, GtfsRealtime.TripDescriptor tripDescriptor) {
-        if (tripDescriptor.getRouteId() != null) {
-            LineRef lineRef = new LineRef();
-            lineRef.setValue(tripDescriptor.getRouteId());
-            monitoredVehicleJourney.setLineRef(lineRef);
-        }
-
-        if (StringUtils.isNotEmpty(tripDescriptor.getStartDate()) && StringUtils.isNotEmpty(tripDescriptor.getStartTime())) {
-            Timestamp.Builder tsBuilder = Timestamp.newBuilder();
-            try {
-                Date date = gtfsRtDateFormat.parse(tripDescriptor.getStartDate() + " " + tripDescriptor.getStartTime());
-                tsBuilder.setSeconds(date.getTime() / 1000);
-                ZonedDateTime startDate = ZonedDateTime.ofInstant(date.toInstant(), ZoneId.systemDefault());
-                monitoredVehicleJourney.setOriginAimedDepartureTime(startDate);
-            } catch (ParseException e) {
-                logger.error("Unable to parse trip descriptor start date/time:" + tripDescriptor.getStartDate() + " " + tripDescriptor.getStartTime());
-            }
-        }
-    }
-
-    private static void mapVehicleRef(VehicleActivityStructure activity, VehicleActivityStructure.MonitoredVehicleJourney monitoredVehicleJourney, GtfsRealtime.VehicleDescriptor vehicleDescriptor) {
-        if (vehicleDescriptor.getId() != null) {
-            VehicleRef vehicleRef = new VehicleRef();
-            vehicleRef.setValue(vehicleDescriptor.getId());
-            monitoredVehicleJourney.setVehicleRef(vehicleRef);
-            VehicleMonitoringRefStructure vehicleRefStruct = new VehicleMonitoringRefStructure();
-            vehicleRefStruct.setValue(vehicleDescriptor.getId());
-            activity.setVehicleMonitoringRef(vehicleRefStruct);
-        }
-    }
-
-    private static void mapLicensePlate(VehicleActivityStructure.MonitoredVehicleJourney monitoredVehicleJourney, GtfsRealtime.VehicleDescriptor vehicleDescriptor) {
-        if (vehicleDescriptor.hasLicensePlate()) {
-            VehicleRef vehicleRef = new VehicleRef();
-            vehicleRef.setValue(vehicleDescriptor.getLicensePlate());
-            monitoredVehicleJourney.setVehicleRef(vehicleRef);
-        }
-    }
-
-    /**
-     * Maps a GTFS-Realtime {@link GtfsRealtime.VehiclePosition} into a {@link VehicleActivityStructure}.
-     * This method extracts relevant vehicle position details, such as trip information, vehicle reference,
-     * location, status, and occupancy, and structures them into a vehicle activity format.
-     *
-     * @param vehiclePosition The GTFS-Realtime {@link GtfsRealtime.VehiclePosition} containing real-time vehicle data.
-     * @param datasetId       dataset.
-     * @param routeIdList     A list of route IDs used to filter relevant vehicle positions.
-     * @return A {@link VehicleActivityStructure} object representing the structured vehicle activity data,
-     * or {@code null} if the vehicle's route ID is not in the provided list.
-     */
-    public VehicleActivityStructure mapVehicleActivityFromVehiclePosition(GtfsRealtime.VehiclePosition vehiclePosition, String datasetId, List<String> routeIdList) {
-        VehicleActivityStructure activity = new VehicleActivityStructure();
-        GtfsRealtime.TripDescriptor tripDescriptor = vehiclePosition.getTrip();
-        if (CollectionUtils.isNotEmpty(routeIdList)) {
-            String routeIdInCache = "";
-            if (tripDescriptor.hasTripId()) {
-                routeIdInCache = stopTimesService.getRouteId(datasetId, tripDescriptor.getTripId()).orElse("");
-            } else if (tripDescriptor.hasRouteId()) {
-                routeIdInCache = stopTimesService.checkIfKnownRouteId(datasetId, tripDescriptor.getRouteId()) ? tripDescriptor.getRouteId() : "";
-            }
-            if (StringUtils.isNotBlank(routeIdInCache) && !routeIdList.contains(routeIdInCache)) {
-                return null;
-            }
-        }
-
-        FramedVehicleJourneyRefStructure framedVehicleJourneyRefBuilder = new FramedVehicleJourneyRefStructure();
-        framedVehicleJourneyRefBuilder.setDatedVehicleJourneyRef(tripDescriptor.getTripId());
-        DataFrameRefStructure dataFrameRef = new DataFrameRefStructure();
-        dataFrameRef.setValue("");
-        framedVehicleJourneyRefBuilder.setDataFrameRef(dataFrameRef);
-
-        VehicleActivityStructure.MonitoredVehicleJourney monitoredVehiclejourney = new VehicleActivityStructure.MonitoredVehicleJourney();
-        monitoredVehiclejourney.setFramedVehicleJourneyRef(framedVehicleJourneyRefBuilder);
-        monitoredVehiclejourney.setDataSource("MOBIITI");
-
-        activity.setValidUntilTime(ZonedDateTime.now().plusMinutes(5));
-
-        mapTridData(monitoredVehiclejourney, vehiclePosition.getTrip());
-
-        if (vehiclePosition.hasVehicle()) {
-            mapVehicleRef(activity, monitoredVehiclejourney, vehiclePosition.getVehicle());
-            mapLicensePlate(monitoredVehiclejourney, vehiclePosition.getVehicle());
-        }
-
-        mapRecordedAtTime(activity, vehiclePosition);
-        mapPosition(monitoredVehiclejourney, vehiclePosition.getPosition());
-        mapOccupancy(monitoredVehiclejourney, vehiclePosition);
-        mapCongestion(monitoredVehiclejourney, vehiclePosition);
-        mapStatus(activity, monitoredVehiclejourney, vehiclePosition);
-
-        activity.setMonitoredVehicleJourney(monitoredVehiclejourney);
-
-        Extensions extensions = new Extensions();
-        List<Object> extensionElements = new java.util.ArrayList<>(Collections.singletonList(extensions.getAnies()));
-
-        if (vehiclePosition.hasPosition()) {
-            GtfsRealtime.Position position = vehiclePosition.getPosition();
-            if (position.hasOdometer()) {
-                Element odometerElement = ElementUtils.createSimpleExtensionElement("odometer", String.valueOf(position.getOdometer()));
-                if (odometerElement != null) {
-                    extensionElements.add(odometerElement);
-                }
-            }
-        }
-
-        if(!extensionElements.isEmpty()) {
-            activity.setExtensions(extensions);
-        }
-
-        return activity;
-    }
-
-
-}
