--- src/main/java/no/rutebanken/anshar/gtfsrt/readers/AlertReader.java
@@ -1,217 +1,0 @@
-package no.rutebanken.anshar.gtfsrt.readers;
-
-import com.google.transit.realtime.GtfsRealtime;
-import no.rutebanken.anshar.api.GtfsRTApi;
-import no.rutebanken.anshar.gtfsrt.mappers.AlertMapper;
-import no.rutebanken.anshar.routes.siri.handlers.inbound.SituationExchangeInbound;
-import no.rutebanken.anshar.subscription.SiriDataType;
-import no.rutebanken.anshar.subscription.SubscriptionManager;
-import no.rutebanken.anshar.subscription.SubscriptionSetup;
-import no.rutebanken.anshar.subscription.helpers.RequestType;
-import org.apache.camel.Produce;
-import org.apache.camel.ProducerTemplate;
-import org.apache.commons.lang3.BooleanUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Component;
-import uk.org.siri.siri21.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import static no.rutebanken.anshar.routes.validation.validators.Constants.GTFSRT_SX_PREFIX;
-
-/**
- * Class to handle and ingest alert data
- * alert (GTFS-RT) = situation exchange (SIRI)
- */
-
-@Component
-public class AlertReader extends AbstractSwallower {
-
-    private static final Logger logger = LoggerFactory.getLogger(AlertReader.class);
-
-    private final SubscriptionManager subscriptionManager;
-
-    private final AlertMapper alertMapper;
-    private final SituationExchangeInbound situationExchangeInbound;
-    @Produce("direct:send.sx.to.realtime.server")
-    protected ProducerTemplate gtfsrtSxProducer;
-
-
-    public AlertReader(SubscriptionManager subscriptionManager, AlertMapper alertMapper, SituationExchangeInbound situationExchangeInbound) {
-        this.subscriptionManager = subscriptionManager;
-        this.alertMapper = alertMapper;
-        this.situationExchangeInbound = situationExchangeInbound;
-        prefix = GTFSRT_SX_PREFIX;
-        dataType = SiriDataType.SITUATION_EXCHANGE;
-        requestType = RequestType.GET_SITUATION_EXCHANGE;
-    }
-
-
-    /**
-     * Processes and ingests GTFS-Realtime alert data, converting it into a structured format
-     * and handling updates, subscriptions, and message dispatching.
-     *
-     * @param gtfsRTApi             parameters of the gtfs rt
-     * @param routeIdList           A list of route IDs used to filter relevant alerts.
-     * @param completeGTFSRTMessage The complete GTFS-Realtime {@link GtfsRealtime.FeedMessage} containing alert data.
-     */
-    public void ingestAlertData(GtfsRTApi gtfsRTApi, List<String> routeIdList, GtfsRealtime.FeedMessage completeGTFSRTMessage, Long inboundTime) {
-        String datasetId = gtfsRTApi.getDatasetId();
-        List<PtSituationElement> situations = buildSituationList(completeGTFSRTMessage, gtfsRTApi, routeIdList);
-        updateParticipantRef(datasetId, situations);
-        List<String> subscriptionList = getSubscriptions(situations);
-        checkAndCreateSubscriptions(subscriptionList, datasetId);
-        buildSiriAndSend(situations, datasetId);
-        if (BooleanUtils.isTrue(gtfsRTApi.getCloseMissingAlerts())) {
-            situationExchangeInbound.closeMissingAlerts(gtfsRTApi.getDatasetId(), situations, inboundTime);
-        }
-
-    }
-
-    private void buildSiriAndSend(List<PtSituationElement> situations, String datasetId) {
-
-        if (situations.isEmpty()) {
-            logger.info("no situations to ingest on dataset : {}", datasetId);
-            return;
-        }
-
-        Siri siri = new Siri();
-        ServiceDelivery serviceDel = new ServiceDelivery();
-        SituationExchangeDeliveryStructure delStruct = new SituationExchangeDeliveryStructure();
-        SituationExchangeDeliveryStructure.Situations sitStruct = new SituationExchangeDeliveryStructure.Situations();
-        sitStruct.getPtSituationElements().addAll(situations);
-        delStruct.setSituations(sitStruct);
-        serviceDel.getSituationExchangeDeliveries().add(delStruct);
-        siri.setServiceDelivery(serviceDel);
-        sendToRealTimeServer(gtfsrtSxProducer, siri, datasetId);
-    }
-
-    private void updateParticipantRef(String datasetId, List<PtSituationElement> situations) {
-        for (PtSituationElement situation : situations) {
-            if (situation.getParticipantRef() == null) {
-                RequestorRef requestorRef = new RequestorRef();
-                requestorRef.setValue("MOBIITI");
-                situation.setParticipantRef(requestorRef);
-
-                SituationSourceStructure sourceStruct = new SituationSourceStructure();
-                NaturalLanguageStringStructure langStruct = new NaturalLanguageStringStructure();
-                langStruct.setLang("FR");
-                langStruct.setValue(datasetId);
-                sourceStruct.setName(langStruct);
-                situation.setSource(sourceStruct);
-            }
-        }
-    }
-
-    /**
-     * Converts a GTFS-Realtime feed message into a list of {@link PtSituationElement} instances.
-     * Each alert in the feed is transformed into a structured situation element.
-     *
-     * @param feedMessage The GTFS-Realtime {@link GtfsRealtime.FeedMessage} containing alert data.
-     * @param gtfsRTApi   parameters of the gtfsrtAPI
-     * @param routeIdList A list of route IDs used to filter relevant alerts.
-     * @return A list of {@link PtSituationElement} objects representing the structured alert data.
-     */
-    public List<PtSituationElement> buildSituationList(GtfsRealtime.FeedMessage feedMessage, GtfsRTApi gtfsRTApi, List<String> routeIdList) {
-        List<PtSituationElement> situtations = new ArrayList<>();
-
-
-        for (GtfsRealtime.FeedEntity feedEntity : feedMessage.getEntityList()) {
-            if (isEmptyAlert(feedEntity.getAlert()) || (hasNoStartAndNoEndDate(feedEntity.getAlert()) && !gtfsRTApi.getGenerateActivePeriod()))
-                continue;
-
-            alertMapper.mapSituationFromAlert(feedEntity, gtfsRTApi, routeIdList).ifPresent(situtations::add);
-
-        }
-        return situtations;
-
-    }
-
-    private boolean isEmptyAlert(GtfsRealtime.Alert alert) {
-
-        return alert == null || alert.getInformedEntityCount() == 0 ||
-                alert.getHeaderText() == null || alert.getHeaderText().getTranslationList().isEmpty() ||
-                alert.getDescriptionText() == null || alert.getDescriptionText().getTranslationList().isEmpty();
-
-    }
-
-    /**
-     * Checks if an alert has no start and no end defined
-     *
-     * @param alert alter to check from
-     * @return true : no start date and no end end is defined at all
-     * false : at least one start or one end is defined in the alert
-     */
-    private boolean hasNoStartAndNoEndDate(GtfsRealtime.Alert alert) {
-        if (alert.getActivePeriodList().isEmpty()) {
-            return true;
-        }
-
-        for (GtfsRealtime.TimeRange timeRange : alert.getActivePeriodList()) {
-            if (timeRange.hasStart() || timeRange.hasEnd()) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-
-    /**
-     * Read all situation messages and build a list of subscriptions that must be checked(or created if not exists)
-     *
-     * @param situations The list of situations
-     * @return The list of subscription ids build by reading the situations
-     */
-    private List<String> getSubscriptions(List<PtSituationElement> situations) {
-        return situations.stream()
-                .map(this::getSituationSubscriptionId)
-                .toList();
-    }
-
-    private String getSituationSubscriptionId(PtSituationElement situation) {
-        StringBuilder key = new StringBuilder();
-
-        if (situation.getSituationNumber() != null) {
-            key.append(situation.getSituationNumber().getValue());
-            key.append(":");
-        }
-
-        if (situation.getParticipantRef() != null) {
-            key.append(situation.getParticipantRef().getValue());
-        }
-
-        return key.isEmpty() ? key.toString() : "GeneralSubsCriptionId";
-    }
-
-
-    /***
-     * Read the list of subscription ids and for each, check if it exists. If not, a new subscription is created
-     * @param subscriptionsList
-     *  The list of subscription ids
-     */
-    private void checkAndCreateSubscriptions(List<String> subscriptionsList, String datasetId) {
-
-        for (String subsId : subscriptionsList) {
-            if (subscriptionManager.isGTFSRTSubscriptionExisting(prefix + datasetId))
-                //A subscription is already existing for this situation. No need to create one
-                continue;
-            createNewSubscription(subsId, datasetId);
-        }
-    }
-
-    /**
-     * Create a new subscription for the id given in parameter
-     *
-     * @param subscriptionId The id for which a subscription must be created
-     */
-    private void createNewSubscription(String subscriptionId, String datasetId) {
-        SubscriptionSetup setup = createStandardSubscription(subscriptionId, datasetId);
-        setup.setSubscriptionId(prefix + datasetId);
-        setup.setName(prefix + datasetId);
-        subscriptionManager.addSubscription(prefix + datasetId, setup);
-        subscriptionManager.addGTFSRTSubscription(prefix + datasetId);
-    }
-
-}
