+++ src/main/resources/application.properties
@@ -195,4 +195,4 @@ anshar.initial.delivery.facility.monitoring.queue.name=activemq:queue:initial.de
 anshar.initial.delivery.situation.exchange.queue.name=activemq:queue:initial.delivery.situation.exchange?disableReplyTo=true
 anshar.initial.delivery.vehicle.monitoring.queue.name=activemq:queue:initial.delivery.vehicle.monitoring?disableReplyTo=true
 vj.mapping.csv=/tmp/lmd/chouette/mobiiti_technique/ineo/vehicleJourneyMapping.csv
-
+anshar.gtfsrt.retrieve.enabled=false
